refactor: mvc 패턴에 맞춰 역할 재분배
