+++ src/main/java/racingcar/Application.java
@@ -1,10 +1,10 @@
 package racingcar;
 
-import racingcar.controller.MainController;
+import racingcar.controller.GameController;
 
 public class Application {
     public static void main(String[] args) {
-        MainController mainController = new MainController();
-        mainController.play();
+        GameController gameController = new GameController();
+        gameController.play();
     }
 }
