--- src/main/java/racingcar/controller/MainController.java
@@ -1,13 +1,0 @@
-package racingcar.controller;
-
-import racingcar.service.GameService;
-
-public class MainController {
-    GameService gameService = new GameService();
-
-    public void play() {
-        gameService.setUpRace();
-        gameService.runRace();
-        gameService.showWinners();
-    }
-}
